+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @web3icons/utils
 
+## 0.3.18
+
+### Patch Changes
+
+- Updated dependencies
+  - @web3icons/common@0.11.17
+
+## 0.3.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @web3icons/common@0.11.16
+
 ## 0.3.16
 
 ### Patch Changes